+++ common/src/main/java/io/github/cotrin8672/ItemIndicator.kt
@@ -1,6 +1,7 @@
 package io.github.cotrin8672
 
 import io.github.cotrin8672.render.BeeOverlayRenderer
+import io.github.cotrin8672.render.EnchantmentIconOverlayRenderer
 import io.github.cotrin8672.render.OverlayRendererRegistry
 import io.github.cotrin8672.render.ShulkerBoxOverlayRenderer
 import io.github.cotrin8672.util.shulkerBoxItemInstanceList
@@ -15,5 +16,6 @@ object ItemIndicator {
         }
         OverlayRendererRegistry.registerItemOverlayRenderer(BeeOverlayRenderer)
         OverlayRendererRegistry.registerItemOverlayRenderer(BeeOverlayRenderer)
+        OverlayRendererRegistry.registerOverlayRenderer(EnchantmentIconOverlayRenderer)
     }
 }