+++ neoforge/src/main/java/io/github/cotrin8672/neoforge/ItemIndicatorNeoForge.kt
@@ -1,9 +1,10 @@
 package io.github.cotrin8672.neoforge
 
 import io.github.cotrin8672.ItemIndicator
+import net.neoforged.api.distmarker.Dist
 import net.neoforged.fml.common.Mod
 
-@Mod(ItemIndicator.MOD_ID)
+@Mod(ItemIndicator.MOD_ID, dist = [Dist.CLIENT])
 class ItemIndicatorNeoForge {
     init {
         ItemIndicator.init()